Rewritten Update - Second Commit
Added help command, coins command, toplist for the highest coins per guild, basic prefix checker, and basic command checker, as well as the storage of command and argument values. Basically, the basic logic of the bot is done. Now all that needs to be done is adding all of the features of the bot and their respective commands. Untested code, so far.